Safari-609.1.11.1/LayoutTests/js/dom/date-negative-setmonth-expected.txt

This test checks for a regression against Date.setMonth fails with negative values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d.valueOf() is new Date(2004, 9, 15).valueOf()
PASS successfullyParsed is true

TEST COMPLETE
